+++ spec/example_app/db/migrate/20160119024340_add_kind_to_customer.rb
@@ -1,0 +1,5 @@
+class AddKindToCustomer < ActiveRecord::Migration
+  def change
+    add_column :customers, :kind, :string, null: false, default: "standard"
+  end
+end